+++ admin_panel/view_users.php
@@ -64,7 +64,7 @@
 									</button>
 								</td>
 								<td>
-									<button class="btn btn-danger">
+									<button class="btn btn-danger" onclick="return confirm('Bạn có chắc chắn muốn xóa?');">
 										<a href="index.php?delete_user=<?php echo $user_id; ?>"><i class="fa fa-trash"></i> Delete</a>
 									</button>
 								</td>
